// upvertice-ui/src/app/services/models/sponsorship-patch-multipart-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Wrapper for multipart Sponsorship patch request
 */
export interface SponsorshipPatchMultipartRequest {

  /**
   * Image file (optional)
   */
  image?: Blob;

  /**
   * Sponsorship patch request data as JSON string
   */
  request?: string;
}